\name{SymmetricMA}
\alias{SymmetricMA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Smoothing a series using moving average. }
\description{
It smooths the data using moving average of specified window size.  
}
\usage{
SymmetricMA(x, w = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Array-CGH data }
  \item{w}{Window size.}
}

\value{
Smoothed version of original data.
}



\author{ A. I. McLeod and M. S. Islam }


\keyword{nonparametric}
