\name{Unique}
\alias{Unique}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Finding the unique values in the original order. }
\description{
  This finds the unique values in the original order.
}
\usage{
Unique(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A vector representing the series }
}

\value{
unique values of the series.
}

\author{ A. I. McLeod and M. S. Islam }


\examples{
a<-c(4,3,2,4,5,1,2,4,2,1,4,5,3)
Unique(a)
}

\keyword{nonparametric}
